+++ src/back/Data/Keeper.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace KeepersTeam\Webtlo\External\Api\V1;
+namespace KeepersTeam\Webtlo\Data;
 
 /** Данные хранителя. */
-final class KeeperData
+final class Keeper
 {
     public function __construct(
         public readonly int    $keeperId,
