+++ src/back/Enum/KeepingPriority.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace KeepersTeam\Webtlo\External\Api\V1;
+namespace KeepersTeam\Webtlo\Enum;
 
 /** Приоритет хранения раздачи. */
 enum KeepingPriority: int
