+++ src/back/External/Api/Actions/ForumTopics.php
@@ -6,11 +6,11 @@ namespace KeepersTeam\Webtlo\External\Api\Actions;
 
 use Generator;
 use GuzzleHttp\Exception\GuzzleException;
-use KeepersTeam\Webtlo\External\Api\V1\ApiError;
-use KeepersTeam\Webtlo\External\Api\V1\ForumTopic;
-use KeepersTeam\Webtlo\External\Api\V1\ForumTopicsResponse;
-use KeepersTeam\Webtlo\External\Api\V1\KeepingPriority;
-use KeepersTeam\Webtlo\External\Api\V1\TorrentStatus;
+use KeepersTeam\Webtlo\Enum\KeepingPriority;
+use KeepersTeam\Webtlo\Enum\TorrentStatus;
+use KeepersTeam\Webtlo\External\Data\ApiError;
+use KeepersTeam\Webtlo\External\Data\ForumTopic;
+use KeepersTeam\Webtlo\External\Data\ForumTopicsResponse;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
