+++ src/back/External/Api/Actions/Processor.php
@@ -8,8 +8,8 @@ use DateTimeImmutable;
 use GuzzleHttp\Exception\GuzzleException;
 use GuzzleHttp\Promise\Promise;
 use JsonException;
-use KeepersTeam\Webtlo\External\Api\V1\ApiError;
 use KeepersTeam\Webtlo\External\Api\V1\TopicSearchMode;
+use KeepersTeam\Webtlo\External\Data\ApiError;
 use KeepersTeam\Webtlo\External\Shared\Validation;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
