+++ src/back/External/Api/V1/KeepersResponse.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace KeepersTeam\Webtlo\External\Api\V1;
 
 use DateTimeImmutable;
+use KeepersTeam\Webtlo\Data\Keeper;
 
 /**
  * Данные всех известных хранителей из API.
@@ -12,15 +13,15 @@ use DateTimeImmutable;
 final class KeepersResponse
 {
     /**
-     * @param DateTimeImmutable      $updateTime дата получения данных
-     * @param array<int, KeeperData> $keepers    ид хранителя => Данные о нём
+     * @param DateTimeImmutable  $updateTime дата получения данных
+     * @param array<int, Keeper> $keepers    ид хранителя => Данные о нём
      */
     public function __construct(
         public readonly DateTimeImmutable $updateTime,
         public readonly array             $keepers,
     ) {}
 
-    public function getKeeperInfo(int $keeperId): ?KeeperData
+    public function getKeeperInfo(int $keeperId): ?Keeper
     {
         return $this->keepers[$keeperId] ?? null;
     }
