+++ src/back/External/ApiReport/Actions/Processor.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace KeepersTeam\Webtlo\External\ApiReport\Actions;
 
 use JsonException;
-use KeepersTeam\Webtlo\External\Api\V1\ApiError;
+use KeepersTeam\Webtlo\External\Data\ApiError;
 use KeepersTeam\Webtlo\External\Shared\Validation;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Log\LoggerInterface;
